Extract Google search plugin into a separate file
